requirements.txt
fastapi
uvicorn
python-multipart
fpdf2
pillow